requirements-faststream.txt
# FastStream зависимости для QIKI DTMP
# Версия: 1.0
# Дата: 2025-08-19

# Core FastStream
faststream[nats]>=0.5.0
faststream[cli]>=0.5.0

# NATS клиент
nats-py>=2.6.0
prometheus-client>=0.20.0

# Тестирование
pytest>=7.4.0
pytest-asyncio

# Другие утилиты
uvicorn>=0.23.0
